Do not allow token exchange when a single client app is involved
Token exchange in IAM is implemenented mainly to provide a mechanism to
allow controlled delegation of privileges across different client
applications. As such, with this commit we introduce a check that
forbids single client token exchanges.

This commit also introduces a token exchange policy cache mechanism that
should improve performance in token exchange at scale by limiting the
connections to the database.