Improving TOTP input screen post-login, fixing Java 17 Docker upgrade issue
Improving the look and feel of the form for inputting a TOTP after logging in with an MFA user. This form will be the basis of step-up authentication and will be navigated to whenever the user is required to verify their identity again.

Will separate the controller for this stage into two distinct paths so that we have one verify "homepage" where the user can choose which factor of authentication to use and then a sub-resource, /verify/authenticator-app for example, which we POST the TOTP to.

Still to add in error checking of TOTP on the verify form.

Also fixing a problem with the test-client's Dockerfile which meant that the Dockerised IAM could not properly start following the upgrade to Java 17.