Separating MFA verification controllers, adding "Back to login" button
Adding a base controller for handling which factor of authentication the user wishes to choose from and making use of a separate controller for handling the specifics of each factor.

Also adding a "Back to login page" button on the verify endpoint which POSTs to /logout to delete all session data securely.