Removing unnecessary database migrations, extending authentication object to contain amr details
The old method of storing details about the user's authentication methods used didn't work, as the information wasn't flushed to the database properly. This was also poor coding practice in that it involved writing temporary authentication data to the database.

This has been improved by extending the UsernamePasswordAuthenticationToken to include a field which has this information. This will become the new expected authentication object whenever the user is authenticating through username and password. Through this, amr details can then be read and added to the id_token returned through OIDC later