Allow to add certificates with the same subjectDn (#624)
Removing unique constraint on issuer_dn column and making subject_dn and issuer_dn case sensitive columns in iam_x509_cert db table, allow to add certificates with the same subject. If the certificate to be added has the same issuer as an already linked certificate, they are considered the same and then the certificate is only updated. Previously, this happened with certificates having the same subject.
Tests updated as a consequence.