Merge commit from fork
This is to fix the security issue where user with OIDC and SAML login can have their token issued without typing in TOTP. This is because the principle for issuring token is the client and so are not restricted by the `PRE_AUTHENTICATED` role assigned to the user as it just check if the user is authenticated. The issue is that OIDC and SAML can't be given a `Authenticated = false` since it violates the contract with spring `ProviderManager`. This caused the mitreID to issue the token despite user not finishing authentication and a under only a partial authentication. Thus, this commit aims to address the above by:
- A check is added to the `createAccessToken` Method so that an exception is triggered when client asking for access token to be issued to a user that contains the role `PRE_AUTHENTICATED`.
- A test that confirms if the subject has the role  `PRE_AUTHENTICATED`  an access token will not be issued and an exception is raised.