Map entity_id of the RP to the client_id
assigned during explicit client registration.
If the OP finds that it already has an existing client registration
for the requesting RP, the registration is invalidated, meaning
the client is deleted and then recreated.

The client_expiration db table has been renamed to
client_federation_metadata containing the client_details_id,
the expiraton and the entity_id.